Implement Module.jsuno.proxy without JS Proxy
The downside of JS Proxy is that at least popular browser consoles like Chrome
and Firefox don't show a Proxy's "supported" properties (even if we would
advertise them via an ownKeys handler trap), but instead concentrate on the
properties of the underlying target object.  That means that there was no way to
get tab-completion in the browser console for all the UNO interface attributes
and methods supported by a proxy'd object.  (See also the discussion starting at
<https://groups.google.com/g/google-chrome-developer-tools/c/0IXya-uvQ7o>
"Console completion of a proxy's properties".)

So, rather than using a JS Proxy to trap every property access of a proxy'd
object, create upfront a prox object that explicitly defines properties for all
the UNO interface attributes and methods supported by the proxy'd object.

(And which requires a dedicated Module.jsuno.sameUnoObject function now, as
proxy objects no longer can be used directly in place of their targets, so
Module.uno.sameUnoObject can no longer be used on proxy objects.)