Make Module.uno_mainPort also available as Module.jsuno.mainPort
...following up on
<https://git.libreoffice.org/core/+/e952204b4010a80775ce63a9986506f16d8d7cb9%5E%21>
"Emscripten: Establish a channel between browser and LO main threads", and
following the general convention of shielding users of jsuno from the underlying
Module.uno... implementation.  Only the newly introduced Module.uno_main Promise
is something that users of jsuno will need to use directly (as it is provided on
the browser's main thread, where Module.jsuno is not available).