No longer support using Module.uno_InOutParam_* in jsuno
Support for this dated back to the early days when jsuno was more of a
transparent, optional wrapper around the Embind-based UNO layer, rather than the
completely opaque layer of its own that it is today.

And that support isn't actually needed (and making use of it in client code
would need massive uses of .delete(), cf. the now-deleted code in
test/smoketest.js that tested that support), and just added complexity in the
implementation, so lets drop it.